Update index.html
I had added a cute little logo thing to your sidebar, which is basically just your favicon image, on line 149. You can edit the size and style and even the image as you wish, I just wanted to give you a start, if a logo like that is something you want to use.

I also changed the color for the light mode separator on line 99. You can change this back if you want, I just preferred the pastel purple over the blue and i wanted to see if you liked it. Of course it's your site, so you can do with it as you wish.

I had mostly decided to make these changes because I was bored, but also because I figured you'd really like to know that there are people who care about your website, at least one person anyways lol.
:3